+++ config/routes.php
@@ -10,12 +10,22 @@
  * @link        https://github.com/mirko-pagliai/cakephp-assets
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
+use Assets\Routing\Middleware\AssetMiddleware;
 use Cake\Routing\RouteBuilder;
 use Cake\Routing\Router;
 
 Router::plugin(ASSETS, ['path' => '/assets'], function (RouteBuilder $routes) {
-    $routes->connect('/:filename', ['controller' => 'Assets', 'action' => 'asset'], [
-        'filename' => '[a-z0-9]+\.(css|js)',
-        'pass' => ['filename'],
-    ]);
+    $routes->registerMiddleware('asset', new AssetMiddleware);
+
+    $routes->get('/:filename', [])
+        ->setPatterns(['filename' => '[\w\d]+\.(css|js)'])
+        ->setPass(['filename'])
+        ->setMiddleware(['asset']);
 });
+
+//Router::plugin(ASSETS, ['path' => '/assets'], function (RouteBuilder $routes) {
+//    $routes->connect('/:filename', ['controller' => 'Assets', 'action' => 'asset'], [
+//        'filename' => '[a-z0-9]+\.(css|js)',
+//        'pass' => ['filename'],
+//    ]);
+//});