+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.x branch
+## 1.3 branch
+### 1.3.0
+* the plugin now uses the `AssetsMiddleware` instead of a controller to
+    "serve" assets. The `AssetsController` has therefore been deleted.
+
 ## 1.2 branch
 ### 1.2.0
 * updated for CakePHP 3.6.